pyproject.toml
[tool.poetry]
name = "factorio-mod-downloader"
version = "2.0.0"
description = "Download Factorio mods from the command line or GUI with full dependency resolution."
authors = ["Vaibhav Vikas <vbhvvikas@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: X11 Applications",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Games/Entertainment",
]
packages = [
    { include = "factorio_mod_downloader", from="src" },
]

[tool.poetry.urls]
Changelog = "https://github.com/vaibhav_vikas/factorio-mod-downloader/releases"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
# Core dependencies
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
selenium = "^4.25.0"
chromedriver-autoinstaller = "^0.6.4"
# GUI dependencies
pillow = "^11.0.0"
customtkinter = "^5.2.2"
ctkmessagebox = "^2.7"
# CLI dependencies
pyyaml = "^6.0.2"
rich = "^13.9.4"
colorama = "^0.4.6"
# Build dependencies
poetry-pyinstaller-plugin = "^1.4.0"

[tool.poetry.scripts]
factorio-mod-downloader = "factorio_mod_downloader.__main__:main"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.1"
pyinstaller = "^6.13.0"
isort = ">=5.10.1"
black = "^24.10.0"

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2  

[tool.black]
line-length = 100
skip-string-normalization = false  

[tool.pylint]
disable = [
    "missing-docstring",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-pyinstaller-plugin]
disable-upx = true
recursive-copy-metadata = [
    "requests",
    "chromedriver_autoinstaller",
    "selenium",
    "customtkinter",
    "rich",
    "colorama"
]

[tool.poetry-pyinstaller-plugin.scripts]
factorio_mod_downloader = { source = "src/factorio_mod_downloader/__main__.py", type = "onefile", bundle = true, add_version = true, icon = "factorio_downloader.ico", windowed = true }

[tool.poetry-pyinstaller-plugin.collect]
all = ['factorio_mod_downloader', 'mod_downloader']

[tool.poetry-pyinstaller-plugin.include]
"factorio_downloader.ico" = "."
